corrections

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -13,5 +13,5 @@ int get_endianness(void)
 	a = 1;
 	b = (char *) &a;
 
-	return ((int)*b);
+	return (*b);
 }
